Switched from floats to doubles on the entire code

+++ includes/BVHNode.hpp
@@ -12,7 +12,7 @@ class   BVHNode : public Hittable
         virtual ~BVHNode(void) = default;
         BVHNode(std::vector<std::shared_ptr<Hittable>> hittables);
         BVHNode(std::vector<std::shared_ptr<Hittable>> hittables, size_t start, size_t end);
-		virtual bool    hit(Ray& ray, float t_max) const override;
+		virtual bool    hit(Ray& ray, double t_max) const override;
 		virtual bool    createBoundingBox(AABB& outputBoundingBox) const override;
 
     private:
