+++ includes/RefractiveIndexes.hpp
@@ -1,19 +1,19 @@
 #ifndef REFRACTIVEINDEXES_HPP
 #define REFRACTIVEINDEXES_HPP
 
-const float RI_GLASS = 1.52f;
-const float RI_WATER = 1.333f;
-const float RI_DIAMOND = 2.417f;
-const float RI_SAPPHIRE = 1.773f;
-const float RI_ICE = 1.31f;
-const float RI_AIR = 1.000293f;
-const float RI_CARBON_DIOXIDE = 1.00045f;
-const float RI_HELIUM = 1.000036f;
-const float RI_HYDROGEN = 1.000132f;
-const float RI_KEROSENE = 1.39f;
-const float RI_ACETONE = 1.36f;
-const float RI_AMBER = 1.55f;
-const float RI_LIQUID_HELIUM = 1.025f;
-const float RI_SILICON = 3.45f;
+const double RI_GLASS = 1.52;
+const double RI_WATER = 1.333;
+const double RI_DIAMOND = 2.417;
+const double RI_SAPPHIRE = 1.773;
+const double RI_ICE = 1.31;
+const double RI_AIR = 1.000293;
+const double RI_CARBON_DIOXIDE = 1.00045;
+const double RI_HELIUM = 1.000036;
+const double RI_HYDROGEN = 1.000132;
+const double RI_KEROSENE = 1.39;
+const double RI_ACETONE = 1.36;
+const double RI_AMBER = 1.55;
+const double RI_LIQUID_HELIUM = 1.025;
+const double RI_SILICON = 3.45;
 
 #endif
