+++ srcs/Vector2.cpp
@@ -7,12 +7,12 @@
 // Constructs the Vector2 with default values
 Vector2::Vector2(void)
 {
-	this->_x = 0.0f;
-	this->_y = 0.0f;
+	this->_x = 0.0;
+	this->_y = 0.0;
 }
 
 // Constructs the Vector2 with custom values
-Vector2::Vector2(float x, float y)
+Vector2::Vector2(double x, double y)
 {
 	this->_x = x;
 	this->_y = y;
